# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Sera",
        "description": "Laravel Sera Test Code",
        "contact": {
            "email": "samuelbenedictx@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/contact": {
            "get": {
                "tags": [
                    "Contact"
                ],
                "summary": "Get list of Contact",
                "operationId": "64e740422f8c163296d8a6ea941e706c",
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "-NDn1_FMcIFJ8Tbl3Xlc": {
                                                        "type": "string",
                                                        "format": "array",
                                                        "example": {
                                                            "email": "example@example.com",
                                                            "name": "example",
                                                            "phone": "088812321321"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Contact"
                ],
                "summary": "Add Contact",
                "operationId": "303f6bf6aec91b25ee8229e29280f716",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "title": {
                                                "type": "string"
                                            },
                                            "email": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "example@example.com",
                                    "name": "example",
                                    "phone": "12324123"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "format": "array",
                                            "example": {
                                                "name": "example",
                                                "email": "example@example.com",
                                                "phone": "12324123"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/customer/contact/{uid}": {
            "get": {
                "tags": [
                    "Contact"
                ],
                "summary": "Get Detail Contact",
                "operationId": "20d3a3f7f69021c7ad226ac3e4fb1c0a",
                "parameters": [
                    {
                        "name": "uid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "format": "array",
                                            "example": {
                                                "name": "example",
                                                "email": "example@example.com",
                                                "phone": "12324123"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/contact/{uid}": {
            "put": {
                "tags": [
                    "Contact"
                ],
                "summary": "Update Contact",
                "operationId": "d607daeb30f42a35b2c00dbfd28d12a9",
                "parameters": [
                    {
                        "name": "uid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "title": {
                                                "type": "string"
                                            },
                                            "email": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "example@example.com",
                                    "name": "example",
                                    "phone": "12324123"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "format": "array",
                                            "example": {
                                                "name": "example",
                                                "email": "example@example.com",
                                                "phone": "12324123"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/contact/{_id}": {
            "delete": {
                "tags": [
                    "Contact"
                ],
                "summary": "Delete Contact",
                "operationId": "93a69077e681f05908fa874b8a930832",
                "parameters": [
                    {
                        "name": "_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contact data deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Contact data deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}